Refactored icedata dataset parsers
