+++ icedata/datasets/birds/parser.py
