+++ icedata/datasets/coco/__init__.py
@@ -1,3 +1,3 @@
 from icedata.datasets.coco.data import *
-from icedata.datasets.coco.parsers import *
+from icedata.datasets.coco.parser import *
 from icedata.datasets.coco import trained_models
