+++ icedata/datasets/fridge/__init__.py
@@ -1,3 +1,3 @@
 from icedata.datasets.fridge.data import *
-from icedata.datasets.fridge.parsers import *
+from icedata.datasets.fridge.parser import *
 from icedata.datasets.fridge import trained_models
