+++ icedata/datasets/pennfudan/__init__.py
@@ -1,3 +1,3 @@
 from icedata.datasets.pennfudan.data import *
-from icedata.datasets.pennfudan.parsers import *
+from icedata.datasets.pennfudan.parser import *
 from icedata.datasets.pennfudan import trained_models
