+++ icedata/datasets/pets/__init__.py
@@ -1,3 +1,3 @@
 from icedata.datasets.pets.data import *
-from icedata.datasets.pets.parsers import *
+from icedata.datasets.pets.parser import *
 from icedata.datasets.pets import trained_models
